requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.5
blinker==1.9.0
build==1.3.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.2
cleo==2.1.0
click==8.3.1
cloudpickle==3.1.2
crashtest==0.4.1
cryptography==46.0.3
dataclasses-json==0.6.7
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
dulwich==0.24.10
fastapi==0.118.3
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-adk==1.20.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.129.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.35.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-modelarmor==0.3.0
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.6.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.53.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.23.2
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.12.0
numpy==2.3.1
openai==1.95.0
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.10.18
packaging==24.2
pandas==2.3.3
pbs-installer==2025.11.20
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.5.0
poetry==2.2.1
poetry-core==2.2.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
PyMuPDF==1.26.4
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.29.0
rsa==4.9.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlalchemy-spanner==1.17.1
sqlparse==0.5.4
sse-starlette==3.0.3
starlette==0.48.0
streamlit==1.51.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
trove-classifiers==2025.11.14.15
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
watchdog==6.0.0
websockets==15.0.1
xattr==1.3.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
